refactor: Add separated service interface and implementation
- NoteService 인터페이스를 생성하여 서비스 계약 정의
- NoteServiceImpl 클래스에서 NoteService 인터페이스를 구현하도록 리팩토링
- NoteServiceImpl의 패키지를 'impl' 하위로 이동하여 구조 개선
- 서비스 레이어의 유연성과 테스트 용이성 향상
