+++ src/main/java/com/fls/animecommunity/animesanctuary/controller/PostController.java
@@ -4,7 +4,7 @@ import com.fls.animecommunity.animesanctuary.model.note.Note;
 import com.fls.animecommunity.animesanctuary.model.note.dto.NoteRequestsDto;
 import com.fls.animecommunity.animesanctuary.model.note.dto.NoteResponseDto;
 import com.fls.animecommunity.animesanctuary.model.note.dto.SuccessResponseDto;
-import com.fls.animecommunity.animesanctuary.service.NoteService;
+import com.fls.animecommunity.animesanctuary.service.NoteServiceImpl;
 
 import lombok.RequiredArgsConstructor;
 
@@ -20,7 +20,7 @@ import java.util.List;
 public class PostController {
 
     //IoC
-    private final NoteService noteService;
+    private final NoteServiceImpl noteService;
     
     //create
     @PostMapping("notes")
